Add user role checks

+++ controller/denuncia.controller.php
@@ -1,15 +1,18 @@
 <?php
 require_once '../model/denunciaDAO.php';
 
-class DenunciaController {
+class DenunciaController
+{
     private $model;
 
-    public function __construct() {
+    public function __construct()
+    {
         $this->model = new DenunciaDAO();
     }
-    public function listarDenuncias() {
-        if (!isset($_SESSION['user_id'] ) || $_SESSION['rol_id'] != '1') {
-            header("Location: index.php?c=User&a=login");
+    public function listarDenuncias()
+    {
+        if (!isset($_SESSION['user_id']) || $_SESSION['rol_id'] != '1') {
+            header("Location: index.php?c=Pais&a=index");
             exit();
         }
 
@@ -19,7 +22,12 @@ class DenunciaController {
         require_once '../view/footer.php';
     }
 
-    public function eliminarDenunciasConContenidoId() {
+    public function eliminarDenunciasConContenidoId()
+    {
+        if (!isset($_SESSION['user_id']) || $_SESSION['rol_id'] != '1') {
+            header("Location: index.php?c=Pais&a=index");
+            exit();
+        }
         $contenidoId = $_POST['contenidoId'];
         $this->model->eliminarDenunciasConContenidoId($contenidoId);
         header("Location: index.php?c=Denuncia&a=listarDenuncias");
